+++ src/app/content/highlights/hooks/loadMore.ts
@@ -1,28 +1,38 @@
-import { GetHighlightsColorsEnum } from '@openstax/highlighter/dist/api';
-import { ActionHookBody } from '../../../types';
+import { GetHighlightsColorsEnum, Highlight } from '@openstax/highlighter/dist/api';
+import { ActionHookBody, AppServices, MiddlewareAPI, Store } from '../../../types';
 import { actionHook } from '../../../utils';
 import { book as bookSelector } from '../../selectors';
 import { loadMoreSummaryHighlights, receiveSummaryHighlights, setSummaryFilters } from '../actions';
-import { summaryPageSize } from '../constants';
+import { maxHighlightsPerPage, summaryPageSize } from '../constants';
 import * as select from '../selectors';
+import { SummaryHighlightsPagination } from '../types';
 import {
-  fetchFunctionBody,
   fetchHighlightsForSource,
+  formatReceivedHighlights,
   getNewSources,
   incrementPage,
-  loadMoreByFunction,
 } from './utils';
 
-export const loadUntilPageSize: fetchFunctionBody = async({
+export const loadUntilPageSize = async({
   previousPagination,
   ...args
-}) => {
+}: {
+  previousPagination: SummaryHighlightsPagination, getState: Store['getState'],
+  highlightClient: AppServices['highlightClient'],
+  highlights?: Highlight[]
+  sourcesFetched: string[],
+  pageSize?: number,
+}): Promise<{pagination: SummaryHighlightsPagination, highlights: Highlight[]}> => {
   const state = args.getState();
   const book = bookSelector(state);
   const {colors} = select.summaryFilters(state);
   const {page, sourceIds, perPage} = previousPagination
     ? incrementPage(previousPagination)
-    : {sourceIds: getNewSources(state, args.sourcesFetched, summaryPageSize), page: 1, perPage: summaryPageSize};
+    : {
+      page: 1,
+      perPage: args.pageSize || maxHighlightsPerPage,
+      sourceIds: getNewSources(state, args.sourcesFetched, args.pageSize),
+    };
 
   if (!book || sourceIds.length === 0) {
     return {pagination: null, highlights: args.highlights || []};
@@ -36,7 +46,7 @@ export const loadUntilPageSize: fetchFunctionBody = async({
     prevHighlights: args.highlights,
   });
 
-  if (highlights.length < summaryPageSize) {
+  if (highlights.length < summaryPageSize || !args.pageSize) {
     return loadUntilPageSize({
       ...args,
       highlights,
@@ -48,9 +58,28 @@ export const loadUntilPageSize: fetchFunctionBody = async({
   return {pagination, highlights};
 };
 
+export const loadMore = async({getState, highlightClient}: MiddlewareAPI & AppServices, pageSize?: number) => {
+  const state = getState();
+  const locationFilters = select.highlightLocationFilters(state);
+  const previousPagination = select.summaryPagination(state);
+  const sourcesFetched = Object.keys(select.loadedCountsPerSource(state));
+
+  const {pagination, highlights} = await loadUntilPageSize({
+    getState,
+    highlightClient,
+    pageSize,
+    previousPagination,
+    sourcesFetched,
+  });
+
+  const formattedHighlights = formatReceivedHighlights(highlights, locationFilters);
+
+  return {formattedHighlights, pagination};
+};
+
 export const hookBody: ActionHookBody<typeof setSummaryFilters | typeof loadMoreSummaryHighlights> =
   (services) => async() => {
-    const {formattedHighlights, pagination} = await loadMoreByFunction(loadUntilPageSize, services);
+    const {formattedHighlights, pagination} = await loadMore(services, summaryPageSize);
     services.dispatch(receiveSummaryHighlights(formattedHighlights, pagination));
   };
