+++ src/app/content/highlights/utils/paginationUtils.ts
@@ -15,6 +15,7 @@ import {
   findTreePages
 } from '../../utils/archiveTreeUtils';
 import { stripIdVersion } from '../../utils/idUtils';
+import { maxResourcesPerFetch } from '../constants';
 import { CountsPerSource, HighlightLocationFilters, SummaryFilters } from '../types';
 
 const totalOfCountsForSource: (counts: CountsPerSource[string]) => number = flow(
@@ -41,7 +42,19 @@ export const getNextPageSources = (
   const pages = findTreePages(tree);
 
   const reduceUntilPageSize = reduceUntil(
-    (counts: CountsPerSource) => nextPageSize ? totalOfCountsPerSource(counts) >= nextPageSize : false
+    (counts: CountsPerSource) => {
+      const reachedResourceLimit = Object.keys(counts).length >= maxResourcesPerFetch;
+
+      if (reachedResourceLimit) {
+        return true;
+      }
+
+      if (nextPageSize && totalOfCountsPerSource(counts) >= nextPageSize) {
+        return true;
+      }
+
+      return false;
+    }
   );
 
   const addPageCount = (counts: CountsPerSource, page: LinkedArchiveTreeSection) => {
